Reply crud, 연관관계 설정

+++ src/main/java/com/example/demo/test/interfaces/reply/dto/ReplyListDto.java
@@ -17,7 +17,7 @@ public class ReplyListDto {
             example = "151",
             required = true
     )
-    public int replySeq;
+    public Long replySeq;
 
     @Schema(
             description = "댓글 내용",
